azure_function/requirements.txt
azure-functions>=1.24.0
pandas>=2.3.3
numpy>=2.3.5
scikit-learn>=1.7.2
azure-storage-blob>=12.27.1